Move Link(Group) types to root Admin/types
