+++ src/pages/Admin/Charity/index.tsx
@@ -1,5 +1,5 @@
 import { Navigate, Route, Routes } from "react-router-dom";
-import { LinkGroup } from "../Sidebar/types";
+import { LinkGroup } from ".././types";
 import { adminRoutes } from "constants/routes";
 import { useAdminResources } from "../Guard";
 import Layout from "../Layout";
