+++ src/pages/Admin/types.ts
@@ -13,6 +13,22 @@ import {
   RegistrarOwnerPayload,
 } from "types/contracts";
 import { DiffSet } from "types/utils";
+import { IconType } from "components/Icon";
+
+export type Link = {
+  title: string;
+  to: string;
+  icon: {
+    type: IconType;
+    size: number;
+  };
+  end?: boolean;
+};
+
+export type LinkGroup = {
+  title?: string;
+  links: Link[];
+};
 
 export type AdminParams = { id: string; type: string /**AccountType */ };
 export type ProposalParams = { id: string };
